split main.py and main.kv out into a few reasonable concerns, makes it easier to reason about each section
